functions/requirements.txt
firebase_functions~=0.1.0
firebase-admin>=6.8.0
google-cloud-firestore>=2.11.0
firebase-functions==0.1.1
numpy>=1.26.2
statistics==1.0.3.5
